requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Flask-Login==0.6.2
Flask-Session==0.4.0
mysql-connector-python==8.1.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
requests==2.31.0
reportlab==4.0.4
